create controllers and route
- create controllers for user and lounge
