+++ app.js
@@ -6,17 +6,17 @@ var logger = require("morgan");
 var mongoose = require("mongoose");
 
 var indexRouter = require("./routes/index");
-var usersRouter = require("./routes/users");
+var userRouter = require("./routes/user");
 
 var app = express();
 
-// database setup
-mongoose.connect("mongodb://localhost/werewolf", { useNewUrlParser: true });
-mongoose.connection
-  .on("error", console.error.bind(console, "connection error:"))
-  .once("open", () => {
-    console.log("Database connected!");
-  });
+// // database setup
+// mongoose.connect("mongodb://localhost/werewolf", { useNewUrlParser: true });
+// mongoose.connection
+//   .on("error", console.error.bind(console, "connection error:"))
+//   .once("open", () => {
+//     console.log("Database connected!");
+//   });
 
 // view engine setup
 app.set("views", path.join(__dirname, "views"));
@@ -29,7 +29,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "public")));
 
 app.use("/", indexRouter);
-app.use("/users", usersRouter);
+app.use("/user", userRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
