+++ views/user/profile.pug
@@ -1,0 +1,1 @@
+h1 profile
