+++ views/user/profile_edit.pug
@@ -1,0 +1,1 @@
+h1 profile_edit
